# public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: SliMail API
  description: |
    API transactionnelle pour l'envoi d'emails et la gestion des automatisations.

    ## Authentification

    Toutes les requêtes API doivent inclure une clé API dans le header `Authorization`:

    ```
    Authorization: Bearer YOUR_API_KEY
    ```

    Les clés API peuvent être générées depuis votre tableau de bord SliMail.

    ## Rate Limiting

    - 100 requêtes par minute pour les endpoints de lecture
    - 50 requêtes par minute pour les endpoints d'envoi

    ## Codes d'erreur

    | Code | Description |
    |------|-------------|
    | 400 | Requête invalide |
    | 401 | Non authentifié |
    | 403 | Non autorisé |
    | 404 | Ressource non trouvée |
    | 422 | Erreur de validation |
    | 429 | Rate limit dépassé |
    | 500 | Erreur serveur |

  version: 1.0.0
  contact:
    name: Support SliMail
    email: support@slimail.com
  license:
    name: Propriétaire
    url: https://slimail.com/terms
servers:
  - url: https://api.slimail.com/v1
    description: Production
  - url: https://sandbox.slimail.com/v1
    description: Sandbox (test)
...
paths:
...
  /automations/{automation}/enroll:
    post:
      tags:
        - Automations
      summary: Inscrire un contact
      description: |
        Inscrit un contact dans une automatisation.

        Si le contact n'existe pas, il sera créé automatiquement.
      operationId: enrollContact
      security:
        - BearerAuth: []
      parameters:
        - name: automation
          in: path
          required: true
          description: ID de l'automatisation
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                first_name:
                  type: string
                last_name:
                  type: string
                custom_fields:
                  type: object
                  additionalProperties: true
            example:
              email: "nouveau@example.com"
              first_name: "Marie"
              last_name: "Martin"
              custom_fields:
                company: "Tech Corp"
                plan: "premium"
      responses:
        '200':
          description: Contact inscrit
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  contact_id:
                    type: integer
                  enrollment_id:
                    type: integer
              example:
                success: true
                message: "Contact inscrit dans l'automatisation"
                contact_id: 789
                enrollment_id: 456
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Clé API à inclure dans le header Authorization
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
...
    Unauthorized:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Clé API invalide ou manquante"
    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "La ressource demandée n'existe pas"
    ValidationError:
      description: Erreur de validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            message: "Les données fournies sont invalides"
            errors:
              to:
                - "Le champ to est obligatoire"
              subject:
                - "Le champ subject est obligatoire"
...